ai-service/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
langchain==0.1.6
langchain-groq==0.0.1
langchain-community==0.0.19
PyPDF2==3.0.1
python-dotenv==1.0.1
pydantic==2.6.1
python-multipart==0.0.9
